requirements.txt
langchain
langchain-community
langgraph
crewai
ollama
beautifulsoup4
requests
aiohttp
pypdf
python-dotenv
ujson
asyncio